requirements.txt
matplotlib==3.5.1
numpy==1.20.1
pandas==1.2.3
pandastable==0.12.2.post1
Pillow==9.1.0
python_ternary==1.0.8
ternary==0.1

